Add visual feedback to show 4 quadrants
